Add files via upload
This is the data for HealthCare.
important_HCdata.mat contains 20 matrixs plus a Date vector (just in case). Each matrix is name as the stock and has 2 columns, firsth the closing price, second the volume.
I tend to think that this is not going to be the way we will manage our data. 
The final way coudl be a single matrix for closing price and another for volume, in which each column is a stock. I will all depend on how we are going to implement our network